requirements.txt
flask>=2.3.0
ultralytics>=8.0.0
opencv-python>=4.8.0
pillow>=9.0.0
numpy>=1.24.0
werkzeug>=2.3.0 